fix: feedback from reviewer and sonarcloud
