+++ src/screens/TopicPageScreen/elements/BottomSheetFollow.tsx
@@ -7,10 +7,8 @@ import dimen from '../../../utils/dimen';
 import {fonts} from '../../../utils/fonts';
 import {COLORS} from '../../../utils/theme';
 import {Button} from '../../../components/Button';
-import IconCheckCircleOutline from '../../../assets/icon/IconCheckCircleOutline';
 import useChatClientHook from '../../../utils/getstream/useChatClientHook';
-import ProfilePicture from '../../ProfileScreen/elements/ProfilePicture';
-import IconIncognito from '../../../assets/icon/IconIncognito';
+import UserItem from './UserItem';
 
 export type Follow = 'signed' | 'incognito' | '';
 
@@ -26,81 +24,7 @@ interface BottomSheetFollowProps {
   getTopicDetail: (topicName: string) => void;
   followType: Follow;
   setFollowType: (followType: Follow) => void;
-  ref: Ref<RBSheet>;
 }
-
-type UserItemProps = {
-  type: 'signed' | 'incognito';
-  username: string;
-  profilePicture: string;
-  isFollowing: boolean;
-  onPress?: () => void;
-};
-
-const UserItem = ({type, username, profilePicture, isFollowing, onPress}: UserItemProps) => {
-  return (
-    <TouchableOpacity
-      onPress={onPress}
-      style={{
-        flexDirection: 'row',
-        alignItems: 'center',
-        paddingVertical: dimen.normalizeDimen(14),
-        paddingHorizontal: dimen.normalizeDimen(20),
-        borderBottomColor: COLORS.gray,
-        borderBottomWidth: 1
-      }}>
-      <View style={{flex: 1, flexDirection: 'column'}}>
-        <View
-          style={{
-            flexDirection: 'row',
-            alignItems: 'center'
-          }}>
-          <View
-            style={{
-              marginRight: 8
-            }}>
-            {type === 'signed' ? (
-              <ProfilePicture
-                size={20}
-                width={0}
-                profilePicPath={profilePicture}
-                anonBackgroundColor=""
-                anonEmojiCode=""
-              />
-            ) : (
-              <IconIncognito color={COLORS.white} />
-            )}
-          </View>
-          <Text
-            style={{
-              fontFamily: fonts.inter[400],
-              fontSize: 14,
-              color: COLORS.black
-            }}>
-            {type === 'signed' ? `as @${username}` : 'Incognito'}
-          </Text>
-        </View>
-        <Text
-          style={{
-            fontFamily: fonts.inter[400],
-            fontSize: 12,
-            color: COLORS.blackgrey,
-            marginTop: 4,
-            lineHeight: 18
-          }}>
-          {type === 'signed'
-            ? 'You’ll be visible as a community member.'
-            : 'You’ll be a hidden member of the community.'}
-        </Text>
-      </View>
-      {isFollowing && (
-        <IconCheckCircleOutline
-          color={type === 'signed' ? COLORS.signed_primary : COLORS.anon_primary}
-        />
-      )}
-    </TouchableOpacity>
-  );
-};
 
 const BottomSheetFollow = forwardRef((props: BottomSheetFollowProps, ref: Ref<RBSheet>) => {
   const {
@@ -115,41 +39,39 @@ const BottomSheetFollow = forwardRef((props: BottomSheetFollowProps, ref: Ref<RB
     onClose
   } = props;
   const {followTopic} = useChatClientHook();
-  const handleFollowTopic = async ({type}: {type: Follow}) => {
+
+  const handleFollowTopic = async ({type: followTypeParam}: {type: Follow}) => {
+    const isIncognito = followTypeParam === 'incognito';
+
+    const followTypeCheck = (targetType: Follow) => followType === targetType || followType === '';
+
+    const follow = async (type: Follow) => {
+      setFollowType(type);
+      setIsFollow(true);
+      try {
+        await followTopic(topicName, isIncognito);
+        getTopicDetail(topicName);
+      } catch (error) {
+        if (__DEV__) {
+          console.log(error);
+        }
+      }
+    };
+
+    onClose();
+
     if (isFollow) {
-      if (followType === '') {
+      if (followTypeCheck('')) {
         setMemberCount(memberCount - 1);
       }
     } else {
       setMemberCount(memberCount + 1);
     }
-    onClose();
-    if (type === 'signed') {
-      if (followType === '' || followType === 'incognito') {
-        setFollowType('signed');
-        setIsFollow(true);
-        try {
-          await followTopic(topicName, false);
-          getTopicDetail(topicName);
-        } catch (error) {
-          if (__DEV__) {
-            console.log(error);
-          }
-        }
-      }
-    } else if (type === 'incognito') {
-      if (followType === '' || followType === 'signed') {
-        setFollowType('incognito');
-        setIsFollow(true);
-        try {
-          await followTopic(topicName, true);
-          getTopicDetail(topicName);
-        } catch (error) {
-          if (__DEV__) {
-            console.log(error);
-          }
-        }
-      }
+
+    if (followTypeParam === 'signed' && followTypeCheck('incognito')) {
+      follow('signed');
+    } else if (followTypeParam === 'incognito' && followTypeCheck('signed')) {
+      follow('incognito');
     }
   };
